SPORTS-54 Add index files to components as per Coding Guidelines
